# docs/api/openapi-v1.yaml
...
openapi: 3.1.0
info:
  title: PigOS API
  description: 'Global Swine Farm Management SaaS — Base API + AI Addon platform.


    이 파일은 라이브 라우트에서 자동 생성됨 (scripts/gen_openapi.py). 수기 편집 금지.

    사람용 연동 계약서: docs/mobile-integration-contract.md


    인증: Bearer JWT (/api/v1/auth/login). access 15분 / refresh 7일.

    멀티테넌트: 모든 농장 데이터는 farm_id로 격리.'
  version: 0.1.0
  contact:
    name: PigOS Support
    url: https://pigos.io
  license:
    name: Proprietary
paths:
...
  /api/v1/farms/{farm_id}/boars:
    get:
      tags:
      - Boars
      summary: List Boars
      operationId: list_boars_api_v1_farms__farm_id__boars_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ACTIVE|CULLED|DEAD|TRANSFERRED
          title: Status
        description: ACTIVE|CULLED|DEAD|TRANSFERRED
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoarResponse'
                title: Response List Boars Api V1 Farms  Farm Id  Boars Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Boars
      summary: Create Boar
      operationId: create_boar_api_v1_farms__farm_id__boars_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoarCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoarResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/boars/{boar_id}:
    get:
      tags:
      - Boars
      summary: Get Boar
      operationId: get_boar_api_v1_farms__farm_id__boars__boar_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: boar_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Boar Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoarResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Boars
      summary: Update Boar
      operationId: update_boar_api_v1_farms__farm_id__boars__boar_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: boar_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Boar Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoarUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoarResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BoarCreate:
      properties:
        ear_tag:
          type: string
          maxLength: 30
          title: Ear Tag
        breed:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Breed
        breed_company:
          anyOf:
          - type: string
            maxLength: 30
          - type: 'null'
          title: Breed Company
        entry_date:
          type: string
          format: date
          title: Entry Date
        entry_type:
          anyOf:
          - type: string
            pattern: ^(PURCHASE|BORN|TRANSFER)$
          - type: 'null'
          title: Entry Type
        semen_quality:
          anyOf:
          - type: string
            pattern: ^(EXCELLENT|GOOD|FAIR|POOR)$
          - type: 'null'
          title: Semen Quality
      type: object
      required:
      - ear_tag
      - entry_date
      title: BoarCreate
    BoarResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        ear_tag:
          type: string
          title: Ear Tag
        breed:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed
        breed_company:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed Company
        status:
          type: string
          title: Status
        entry_date:
          type: string
          format: date-time
          title: Entry Date
        entry_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Entry Type
        semen_quality:
          anyOf:
          - type: string
          - type: 'null'
          title: Semen Quality
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - farm_id
      - ear_tag
      - breed
      - breed_company
      - status
      - entry_date
      - entry_type
      - semen_quality
      - created_at
      title: BoarResponse
    BoarUpdate:
      properties:
        ear_tag:
          anyOf:
          - type: string
            maxLength: 30
          - type: 'null'
          title: Ear Tag
        breed:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Breed
        breed_company:
          anyOf:
          - type: string
            maxLength: 30
          - type: 'null'
          title: Breed Company
        status:
          anyOf:
          - type: string
            pattern: ^(ACTIVE|CULLED|DEAD|TRANSFERRED)$
          - type: 'null'
          title: Status
        semen_quality:
          anyOf:
          - type: string
            pattern: ^(EXCELLENT|GOOD|FAIR|POOR)$
          - type: 'null'
          title: Semen Quality
      type: object
      title: BoarUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://api.pigos.io
  description: Production
- url: http://localhost:8000
  description: Local dev (iOS 시뮬)
- url: http://10.0.2.2:8000
  description: Local dev (Android 에뮬)